INGM-689 use error queues, remove same length map
